chore(support): extract shared lib for property scripts
Move the code shared by scrub_person_properties and prune_property_definitions
into a common products/support/scripts/lib package: the retry/backoff HTTP
client, browser-session auth, region host resolution, terminal-safe output, and
the status-code report helper. prune now imports from it; behavior is unchanged.

Generated-By: PostHog Code
Task-Id: 3f15054d-fdee-4520-9efa-c181caca3f73
